+++ src/c_lib/serializer/items.hpp
@@ -5,7 +5,7 @@
 namespace serializer
 {
 
-extern uint32_t item_global_id;
+extern int64_t item_global_id;
 
 void save_items();
 SerializerError save_item(ItemID id);
@@ -14,7 +14,7 @@ SerializerError load_item(ItemID id);
 SerializerError load_item_global_id();
 SerializerError write_item_global_id();
 
-uint32_t get_new_item_global_id();
+int64_t get_new_item_global_id();
 
 void init_item_serializer();
 void teardown_item_serializer();
